feat: 이미지 압축 함수에 옵션값 인자로 받아 재사용할 수 있게 변경
